♻️ Adjust `invariant` Test Configuration (#117)
### 🕓 Changelog

Foundry changed the default value for the `invariant` `depth` parameter
to `500` here: https://github.com/foundry-rs/foundry/pull/7957. Given
the current setup, this would lead to overall 2,500,000 executed calls
(= 5,000 `runs` * 500 `depth`), which would cause the CI to timeout (see
https://github.com/pcaversaccio/createx/actions/runs/9304062368/job/25622386435
as an example). We patch this by setting the new `invariant` `runs`
parameter to 500 and the `depth` parameter to 500, leading to overall
(only) 250,000 (= 500 * 500) calls.

---------

Signed-off-by: Pascal Marco Caversaccio <pascal.caversaccio@hotmail.ch>
Co-authored-by: Matt Solomon <matt@mattsolomon.dev>